requirements.txt
alembic>=1.17.2
asyncpg>=0.31.0
fastapi>=0.128.0
pydantic>=2.12.5
python-dotenv>=1.2.1
pydantic-settings>=2.12.0
sqlalchemy>=2.0.45
uvicorn>=0.40.0